Bugfix INC/DEC (HL)
